requirements.txt
wagtail==7.3.1
psycopg[binary]>=3.2,<4
django-environ>=0.12,<1
whitenoise>=6.9,<7
gunicorn>=23,<26
